+++ kotlin_first_program/src/main_kotlin.kt
@@ -1,8 +1,4 @@
-package com.first.kotlin
 
-/**
- * Created by Karol on 2015-09-12.
- */
 fun main(args: Array<String>) {
     println("czolgaj sie piesku")
     println("na co czekasz?")
